+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Only create `circleci.com/project-slug` annotation on component if there is a `.circleci/config.yml` file.
+
 ## [0.0.6] - 2023-07-27
 
 ### Added