+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.7] - 2023-08-02
+
 ### Added
 
 - Add `backstage.io/techdocs-ref` annotation for components (only if `/README.md` is present).
@@ -50,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial code
 
-[Unreleased]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.0.6...HEAD
+[Unreleased]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.0.7...HEAD
+[0.0.7]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.0.3...v0.0.4